+++ src/main/kotlin/dev/nyon/bbm/config/ConfigScreen.kt
@@ -7,7 +7,7 @@ import net.minecraft.client.gui.screens.Screen
 fun generateYaclScreen(parent: Screen?): Screen = YetAnotherConfigLib("bbm") {
     val general by categories.registering {
         val stepHeight by rootOptions.registering {
-            binding(1f, { config.stepHeight }, { config.stepHeight = it})
+            binding(1f, { config.stepHeight }, { config.stepHeight = it })
             controller = numberField(0f)
             descriptionBuilder {
                 addDefaultText(1)
@@ -15,7 +15,7 @@ fun generateYaclScreen(parent: Screen?): Screen = YetAnotherConfigLib("bbm") {
         }
 
         val playerEjectTicks by rootOptions.registering {
-            binding(0.2f, { config.playerEjectTicks }, { config.playerEjectTicks = it})
+            binding(0.2f, { config.playerEjectTicks }, { config.playerEjectTicks = it })
             controller = numberField(0f, 10000f)
             descriptionBuilder {
                 addDefaultText(1)
@@ -57,6 +57,14 @@ fun generateYaclScreen(parent: Screen?): Screen = YetAnotherConfigLib("bbm") {
         val onlyForPlayers by rootOptions.registering {
             binding(true, { config.onlyForPlayers }, { config.onlyForPlayers = it })
             controller = tickBox()
+            descriptionBuilder {
+                addDefaultText(1)
+            }
+        }
+
+        val extraCollisionDetectionRange by rootOptions.registering {
+            binding(0.0, { config.extraCollisionDetectionRange }, { config.extraCollisionDetectionRange = it })
+            controller = numberField(0.0)
             descriptionBuilder {
                 addDefaultText(1)
             }